moved S3RegistryStore to aws
